+++ src/Products/PloneMeeting/events.py
@@ -100,6 +100,8 @@ def do(action, event):
         event.object.notifyModified()
     elif objectType == 'MeetingAdvice':
         _addManagedPermissions(event.object)
+        # update modification date upon state change
+        notifyModifiedAndReindex(event.object)
 
 
 def onItemTransition(item, event):