+++ CHANGES.rst
@@ -44,6 +44,9 @@ Changelog
   attendee if it was redefined on items (itemAbsent, itemExcused,
   itemSignatories, itemNonAttendees).
   [gbastien]
+- Manage down/up WF for some specific advices so icon `waiting_advices_from.png`
+  is red when down WF, green when up WF again and blue otherwise.
+  [gbastien]
 
 4.2b2 (2020-09-10)
 ------------------